refactor: address copilot reviews

+++ app/eventyay/eventyay_common/views/event.py
@@ -41,7 +41,6 @@ from eventyay.base.services import tickets
 from eventyay.base.settings import DEFAULTS, SETTINGS_AFFECTING_CSS, is_event_series_creation_enabled
 from eventyay.presale.style import regenerate_css
 from eventyay.common.text.path import resolve_media_path
-from eventyay.common.urls import get_file_url_path
 from eventyay.base.services.quotas import QuotaAvailability
 from eventyay.control.forms.event import EventWizardBasicsForm, EventWizardCopyForm, EventWizardFoundationForm
 from eventyay.control.forms.filter import EventFilterForm
@@ -712,7 +711,7 @@ class EventUpdate(
                 current_value = request.event.settings.get(setting_key, as_type=str)
                 if current_value:
                     current_file = resolve_media_path(current_value)
-                    if current_file:
+                    if current_file and not str(current_file).startswith(('http://', 'https://')):
                         default_storage.delete(current_file)
                         base_path, unused_ext = os.path.splitext(current_file)
                         orig_ext = request.event.settings.get(f'{setting_key}_original_ext', as_type=str)
